+++ python/rmm/tests/test_rmm.py
@@ -114,7 +114,7 @@ def test_rmm_device_buffer(size):
     assert len(s) == len(b)
 
     # Test conversion from bytes
-    b2 = rmm.DeviceBuffer.frombytes(s)
+    b2 = rmm.DeviceBuffer.to_device(s)
     assert isinstance(b2, rmm.DeviceBuffer)
     assert len(b2) == len(s)
 
@@ -122,6 +122,35 @@ def test_rmm_device_buffer(size):
     b.resize(2)
     assert b.size == 2
     assert b.capacity() >= b.size
+
+
+@pytest.mark.parametrize(
+    "hb",
+    [
+        b"abc",
+        bytearray(b"abc"),
+        memoryview(b"abc"),
+        np.asarray(memoryview(b"abc")),
+        np.arange(3, dtype="u1"),
+    ],
+)
+def test_rmm_device_buffer_memoryview_roundtrip(hb):
+    mv = memoryview(hb)
+    db = rmm.DeviceBuffer.to_device(hb)
+    hb2 = db.copy_to_host()
+    assert isinstance(hb2, np.ndarray)
+    mv2 = memoryview(hb2)
+    assert mv == mv2
+    hb3a = bytearray(mv.nbytes)
+    hb3b = db.copy_to_host(hb3a)
+    assert hb3a is hb3b
+    mv3 = memoryview(hb3b)
+    assert mv == mv3
+    hb4a = np.empty_like(mv)
+    hb4b = db.copy_to_host(hb4a)
+    assert hb4a is hb4b
+    mv4 = memoryview(hb4b)
+    assert mv == mv4
 
 
 @pytest.mark.parametrize(
@@ -146,19 +175,19 @@ def test_rmm_device_buffer_bytes_roundtrip(hb):
         mv = memoryview(hb)
     except TypeError:
         with pytest.raises(TypeError):
-            rmm.DeviceBuffer.frombytes(hb)
+            rmm.DeviceBuffer.to_device(hb)
     else:
         if mv.format != "B":
             with pytest.raises(ValueError):
-                rmm.DeviceBuffer.frombytes(hb)
+                rmm.DeviceBuffer.to_device(hb)
         elif len(mv.strides) != 1:
             with pytest.raises(ValueError):
-                rmm.DeviceBuffer.frombytes(hb)
+                rmm.DeviceBuffer.to_device(hb)
         elif mv.strides[0] != 1:
             with pytest.raises(ValueError):
-                rmm.DeviceBuffer.frombytes(hb)
+                rmm.DeviceBuffer.to_device(hb)
         else:
-            db = rmm.DeviceBuffer.frombytes(hb)
+            db = rmm.DeviceBuffer.to_device(hb)
             hb2 = db.tobytes()
             mv2 = memoryview(hb2)
             assert mv == mv2
@@ -169,7 +198,7 @@ def test_rmm_device_buffer_bytes_roundtrip(hb):
 
 @pytest.mark.parametrize("hb", [b"", b"123", b"abc"])
 def test_rmm_device_buffer_pickle_roundtrip(hb):
-    db = rmm.DeviceBuffer.frombytes(hb)
+    db = rmm.DeviceBuffer.to_device(hb)
     pb = pickle.dumps(db)
     del db
     db2 = pickle.loads(pb)