+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - PR #261 Add `__bytes__` method to `DeviceBuffer`
 - PR #262 Moved device memory resource files to `mr/device` directory
 - PR #266 Drop `rmm.auto_device`
+- PR #268 Add Cython/Python `copy_to_host` and `to_device`
 
 ## Improvements